refactor(picker): modularize the marks-picker
This can be used to extend and override the key-mappings inside
the `attach_mappings`-function, for example:

```lua
local themes = require("telescope.themes")
local hm_actions = require("telescope._extensions.harpoon_marks.actions")
require("telescope").extensions.harpoon.marks(themes.get_dropdown({
    previewer = false,
    layout_config = { width = 0.6 },
    path_display = { truncate = 10 },
    attach_mappings = function(_, map)
        map("i", "<c-d>", hm_actions.delete_mark_selections)
        map("n", "<c-d>", hm_actions.delete_mark_selections)
        return true
    end,
}))
```
