+++ lua/telescope/_extensions/harpoon.lua
@@ -1,11 +1,11 @@
 local has_telescope, telescope = pcall(require, "telescope")
 
 if not has_telescope then
-    error("harpoon.nvim requires nvim-telescope/telescope.nvim")
+    error("This extension requires telescope.nvim (https://github.com/nvim-telescope/telescope.nvim)")
 end
 
 return telescope.register_extension({
     exports = {
-        marks = require("telescope._extensions.marks"),
+        marks = require("telescope._extensions.harpoon_marks.marks"),
     },
 })
